requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
python-dotenv
cryptography
